UPSTREAM: arm64/efi: handle potential failure to remap memory map
When remapping the UEFI memory map using ioremap_cache(), we
have to deal with potential failure. Note that, even if the
common case is for ioremap_cache() to return the existing linear
mapping of the memory map, we cannot rely on that to be always the
case, e.g., in the presence of a mem= kernel parameter.

At the same time, remove a stale comment and move the memmap code
together.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Acked-by: Mark Rutland <mark.rutland@arm.com>
Acked-by: Mark Salter <msalter@redhat.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>

Bug: 31432001
Change-Id: Ia501b22330dfcd273c0e9a6430072515dcf0c6d4
(cherry picked from commit 7fe5d2b1daf6fd82857a8d0ee47547d7852ebe7b)
Signed-off-by: Sami Tolvanen <samitolvanen@google.com>